feat: Implement data caching in developement mode to make developement easier.
To take advantage of this feature, serve the project by running the `serve.cmd` batch script.
All expensive data fetches should be wrapped in the cache.js function, which accepts the function that gets the data, a file name to save the cache as, and an array of arguments to be passed to the function.
In order to clear the cache pass the flag `/clear` to the serve script.